requirements.txt
certifi==2022.5.18.1
charset-normalizer==2.0.12
idna==3.3
PyMySQL==1.0.2
python-dotenv==0.20.0
requests==2.28.0
urllib3==1.26.9
